Initial implementation of the transition receiver
This currently hardcodes the transition but (with the changes to use this
plugin) seems to generate notifications correctly on both iOS and android.
On android, unlike the local broadcast manager, it generates notifications even
when the app has been force killed. On iOS, we assume that the app will not be
force killed because we prompt users to re-launch the app if they kill it.