Fix the first point retrieval to be more robust
In
https://github.com/e-mission/e-mission-transition-notify/pull/2#issuecomment-267503483,
we encountered an issue with the data/metadata mismatch that caused our assert to fail.
But then if the assert failed, we might actually be returning the wrong point.

Also, the code was slightly different on iOS versus android, since we could
find the transition before the start on android fairly easily but on iOS we had
to go back to the beginning of time.

With this, on both platforms, we search in the 10 minute range around the start
transition and return the last point before the transition (preferable).
If there are points before the transition, we return the first point after the
transition.
We fall back to the first point if there are no points in that range.