Fix issue notifier invoked before FSM
The currently expected flow is:
- FSM is invoked
    - we store the transition
- notifier invoked
    - we assert that the last transition was TRIP_END_DETECTED

While testing with the iPhone7 simulator, however, I discovered that sometimes,
probably due to races in initializing the plugins, the invocation order is
reversed. In other words, the notifier is invoked before the FSM.

Then, the last transition was not TRIP_END_DETECTED, so the assert triggers and
nothing else (trip end, data pushed) happens.

I am not sure how often this happens, but:
- if it fails, the consequences are pretty bad
- we don't actually use the endTransition for anything

So let's replace the assert by an if check for now