Support editing as well
In particular, store the display config, and the edited display config.
And just to be efficient, a list of toggled entries. Note that this is not
strictly necessary since we can reconstruct it by comparing the original and
edited configs.

We need to actually save this in local storage, but let's do that through
a wrapper or directly in native code.