Final javascript + HTML changes for muting/unmuting
Instead of pulling out fields from the config, pass it around directly
so that we can pass it in to the native call.

This is also more principled and easier to extend if you want to show
other fields of the config.

Deal with the fallout of that decision, including modifying the HTML to match.

Tested in the simulator - it works.